+++ service/resource/dnsrecord/create.go
@@ -1,0 +1,27 @@
+package dnsrecord
+
+import (
+	"context"
+
+	"github.com/giantswarm/azuretpr"
+	"github.com/giantswarm/microerror"
+)
+
+// ApplyCreateChange is never called. We do not like it. It is not idempotent.
+func (r *Resource) ApplyCreateChange(ctx context.Context, obj, change interface{}) error {
+	o, err := toCustomObject(obj)
+	if err != nil {
+		return microerror.Maskf(err, "creating DNS records in host cluster")
+	}
+
+	c, err := toDNSRecords(change)
+	if err != nil {
+		return microerror.Maskf(err, "creating DNS records in host cluster")
+	}
+
+	return r.applyCreateChange(ctx, o, c)
+}
+
+func (r *Resource) applyCreateChange(ctx context.Context, obj azuretpr.CustomObject, change dnsRecords) error {
+	return nil
+}
